requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytest>=8.0.0
